python/requirements.txt
adal==1.2.2
autopep8==1.5
azure-common==1.1.25
azure-functions==1.2.0
azure-servicebus==0.50.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
idna==2.9
isodate==0.6.0
msrest==0.6.11
msrestazure==0.6.3
oauthlib==3.1.0
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
PySocks==1.7.1
python-dateutil==2.8.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
tweepy==3.8.0
uamqp==1.2.6
urllib3==1.25.8
